third_party/WebKit/LayoutTests/editing/inserting/insert-div-026.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div class="explanation">
<div class="scenario">
Tests: 
<br>
Inserting blocks for paragraphs should do a better job of finding a block to insert after.
<a href="rdar://problem/3996605">&lt;rdar://problem/3996605&gt;</a> Insert paragraph command puts new block in wrong place, creating difficult-to-handle HTML
</div>
<div class="expected-results">
Expected Results:
<br>
Should see this content in the red box below: <b>fo</b>x
</div>
</div>

<div contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">
<div id="sample" class="editing">fo</div>
<div id="log"></div>
<script>
test(function() {
    var selection = window.getSelection();
    var sample = document.getElementById('sample');

    selection.collapse(sample.firstChild, 0);
    selection.modify('extend', 'forward', 'line');
    document.execCommand('bold');
    selection.modify('move', 'forward', 'character');
    document.execCommand('insertParagraph');
    document.execCommand('bold');
    document.execCommand('InsertText', false, 'x')
    selection.modify('move', 'backward', 'character');
    document.execCommand('delete');

    assert_equals(sample.innerHTML, '<b>fo</b>x');
});
</script>
